cleanup and added unit test for processing step
