--- tests/unit_tests/test_job_pool.py
@@ -1,19 +1,0 @@
-import unittest
-
-from oktoberfest.utils.multiprocessing_pool import JobPool
-
-
-def add_one(i: int):
-    """Test function for multiprocessing pool."""
-    return i + 1
-
-
-class TestJobPool(unittest.TestCase):
-    """Test the JobPool class."""
-
-    def test_jobpool(self):
-        """Unit test for starting and joining multiprocessing pool."""
-        pool = JobPool(2)
-        for i in range(5):
-            pool.apply_async(add_one, [i])
-        pool.check_pool()
